Move `ReshardingState` from `collection/resharding` to `shards/resharding`
